Reorganize file structure and fix compiler warning
